lex: fix segfault on %Start-condition rules on 64-bit hosts
<cond>pattern rules build an RSCON tree node whose right[] slot holds a
real pointer (the list of applicable start-condition numbers), already
cast to the pointer-wide uintptr at the parser.y call site. But
left[]/right[] were declared plain int*, so storing that pointer
truncated it to 32 bits on this 64-bit host. first() (sub2.c) later
reinterprets right[v] as that pointer and dereferences the truncated
address, segfaulting deep in NFA-construction recursion whenever it
reaches such a node.

Found via principia-softwarica's typesetting/pic/picl.lx and
typesetting/grap/grapl.lx, both of which use %Start conditions.

Widen left[]/right[] to uintptr* so the pointer round-trips intact, and
add a regression test with a 4-line minimal repro.

+++ generators/lex/globals.c
@@ -105,11 +105,15 @@ int fptr;
 /*s: global name */
 int *name;
 /*e: global name */
+/* claude: was `int *`; widened to uintptr since right[] must also hold
+ * a real pointer for RSCON nodes -- see the comment above mn2() in
+ * sub1.c for the segfault this used to cause.
+ */
 /*s: global left */
-int *left;
+uintptr *left;
 /*e: global left */
 /*s: global right */
-int *right;
+uintptr *right;
 /*e: global right */
 /*s: global parent */
 int *parent;
